backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
httpx==0.26.0
openai==1.10.0
websockets==12.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
amqp==5.2.0
aiokafka==0.10.0
prometheus-client==0.19.0
